RAM usage optimizations added
The problem of excessive use of RAM in previous versions of Explora UAN is due to the bad practice of Audio playback. The "Decompress On Load" property parameters assigned by Unity by default were assigned, which caused excessive RAM consumption when loading the camera scene.

To solve this problem, the "Streaming" parameter in the properties of the Audio Clip inspector was assigned to all audios.

More information and documentation about audio optimization in Unity at the following link:
https://docs.unity3d.com/Manual/class-AudioClip.html